requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.9
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-extensions==3.2.3
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
graphviz==0.20.1
iniconfig==2.0.0
ipython==8.20.0
jedi==0.19.1
kombu==5.3.5
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.4.4
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.0
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.9
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-extensions==3.2.3
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
graphviz==0.20.1
iniconfig==2.0.0
ipython==8.20.0
jedi==0.19.1
kombu==5.3.5
matplotlib-inline==0.1.6
packaging==23.2
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.4.4
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
